move tests and benches into a sub-crate
also, generalize main tests to all alignments
