+++ Cargo.toml
@@ -7,14 +7,14 @@ description = "no_std compatible pattern scan library using simd"
 repository = "https://github.com/greaka/patterns"
 license = "MIT/Apache-2.0"
 
-[[bench]]
-name = "main"
-harness = false
+# Tests and benchmarks are in a sub crate
+[lib]
+bench = false
+test = false
+doctest = false
 
-[dev-dependencies]
-aligned-vec = "0.5.0"
-criterion = "0.4.0"
-xxhash-rust = { version = "0.8.6", features = ["xxh3"] }
+[workspace]
+members = [ "./tests" ]
 
 [features]
 second_byte = []
